pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "standalone-stt-module"
version = "0.1.0"
description = "Standalone modular Speech-to-Text experimentation pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"numpy>=1.26.0",
	"scipy>=1.11.0",
	"soundfile>=0.12.1",
	"noisereduce>=3.0.0",
	"webrtcvad-wheels>=2.0.14",
	"faster-whisper>=1.0.3"
]

[project.optional-dependencies]
dev = [
	"pytest>=8.0.0"
]

[project.scripts]
stt-module = "stt_module.cli:main"

[tool.setuptools.packages.find]
include = ["stt_module*"]
